+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 ## Unreleased
 
+## [1.2.0](../../compare/1.1.0..1.2.0) - XXXX-XX-XX
+### Added
+- Metamethods `__mul()`, `__call()`, `__index()` and `__newindex()`
+### Changed
+- Replaced comments in source code with doc ones
+
 ## [1.1.0](../../compare/1.0.0..1.1.0) - 2021-12-01
 ### Added
 - `escpattern` and `unescpattern` to replace existing `escregex` and `unescregex` since term "pattern" fits more than "regex" in Lua